+++ src/client/components/UserPage/UserLinkTable/ToolBar/Buttons/CreateLinkButton.jsx
@@ -4,7 +4,7 @@ import { createStyles, makeStyles } from '@material-ui/core'
 
 import userActions from '../../../../../actions/user'
 import ContainedIconButton from './templates/ContainedIconButton'
-import useMinifiedActions from '../util/minifiedActions'
+import useMinifiedActions from '../../../CreateUrlModal/helpers/minifiedActions'
 import OvalContainedButton from './templates/OvalContainedButton'
 
 const useStyles = makeStyles((theme) =>
